wip: Add option to control DER vs PEM encoding, implement for keys.
This is an experiment and it seemed like a good idea. Then I got to the
part where we have to read keys back from disk to generate certs. This
requires we load keys which, using the current approach to key
management, requires we know the encoding.

We either need to implement an approach to managing keys that's
independent of the encoding (a single key file). Or stop storing them in
files. The latter is preferred but will be a significant departure.